commented out --copy-unsafe-links
I will have to figure out how I can handle "symlink has no referent"
issues that occur when a broken symlink is encountered. Perhaps I should
write a script to clean up broken symlinks once in a while, but that
doesn't sound like the right solution because it mutates the source. The
best solutions would be to report unresolvable symlinks and copy the
broken link as is. Currently commenting out the --copy-unsafe-links will
refrain from copying any symlinks but at least also the broken symlinks
will be in the backup. Not sure if this has any merit though.